- Add vector transform patterns needed to get better performance.

Baseline:
```
--------------------------------------------------------------------------------------------
Benchmark                                                  Time             CPU   Iterations
--------------------------------------------------------------------------------------------
BM_mmt4d_384x384x512_8x1x8/process_time/real_time       7.54 ms         7.52 ms           93
```
This PR:
```
--------------------------------------------------------------------------------------------
Benchmark                                                  Time             CPU   Iterations
--------------------------------------------------------------------------------------------
BM_mmt4d_384x384x512_8x1x8/process_time/real_time       3.66 ms         3.65 ms          191
```
Depends on https://reviews.llvm.org/D111561