@asaadaldien 2021-11-04
Running `iree/test/microbenchmarks/linalg_mmt4d.mlir` on Pixel4 core-7
Baseline:
```
--------------------------------------------------------------------------------------------
Benchmark                                                  Time             CPU   Iterations
--------------------------------------------------------------------------------------------
BM_matmul_384x384x512/process_time/real_time            6.64 ms         6.63 ms          106
BM_mmt4d_384x384x512_4x1x4/process_time/real_time       9.74 ms         9.72 ms           72
BM_mmt4d_384x384x512_8x1x8/process_time/real_time       7.10 ms         7.07 ms          100
```
This PR:
```
--------------------------------------------------------------------------------------------
Benchmark                                                  Time             CPU   Iterations
--------------------------------------------------------------------------------------------
BM_matmul_384x384x512/process_time/real_time            6.24 ms         6.22 ms          112
BM_mmt4d_384x384x512_4x1x4/process_time/real_time       4.41 ms         4.40 ms          159
BM_mmt4d_384x384x512_8x1x8/process_time/real_time       3.83 ms         3.82 ms          183
```