@asaadaldien 2021-11-04
> Add lit test for the configuration being set correctly?

done!